Support comparisons including an orphan commit
Handle revision 0 and the null git tree so the resulting diff will display the
results as if those arguments were passed directly to svn diff and git diff,
respectively.

Additionally, automatically detect if the supplied commit (or revision) is an
orphan.  If it is, compare the commit against the null tree so it displays as
expected.

Issue: #14