Retain filename prefixes when using --git
In addition to using Git's file annotations, SVN adopted Git's "a/" and "b/"
prefixes before the src and dst filenames, respectively, when the --git option
is given to svn diff.  They also adopted the use of "/dev/null" when a file
does not exist in the repository for a given version.

Retain these filename prefixes/values from git when the --git option is passed
to git-svn-diff.

Issue: #32