Convert Git symlink diffs to SVN symlink diffs
The contents of a symlink are prefixed with the phrase "link " in SVN as these
files are not actually marked as links (they are marked as "special" in the
metadata).  Add this phrase to support symlink contents (as SVN sees it).

This change does not fix the SVN properties reporting for symlinks.

Issue: #10