Support use of --relative-repo in subdirectory
The output of "git svn info" already takes into account the subdirectory in
which it's run and includes that in the resulting path.  As a result, running
git-svn-diff in a subdirectory with --relative-repo includes the subdirectory
in the file path twice.  Fix this issue by handling the source git-diff with a
relative path, to work well with the relative path given by "git svn info".

Notably, the behavior remains undefined when the --relative-repo and
--relative options are given together.

Issue: #7