Find branch revision corresponding to the one given
In the SVN world, you can supply any revision to the diff tool, irrespective
of whether the current branch experienced changes.  If a branch is changed on
revisions 2 & 5, then specifying a revision of 2, 3, or 4 to the diff tool
will produce the same result.  So, if a revision is supplied, we can use the
lowest revision, 2, for comparison.  The git-svn-diff tool already does this;
however, it uses commit times to quickly find the latest revision on the
current branch, without going over the revision supplied.  This worked well,
unless 2 commits happened within the same second; then it could select a
revision larger than the one supplied.  Add a second check to walk the tree
and ensure the revision used is less than (or equal to) the revision supplied.

Issue: #28