Avoid wildcard expansion in test on Windows
The Windows version of SVN will expand wildcards, so '*' becomes the list of
files in the directory, despite passing a literal '*' in the arguments list.
This probably happens because, on Windows, wildcard expansion is not done by
the command line; it is the responsibility of the application to handle these
characters specially.  And it appears the Windows SVN client always expands
wildcards, even when the argument is not for a file.  Avoid this by using a
value that cannot be misinterpreted.

Issue: #37