Organize main code path into subroutines
Move all of the "global" code into subroutines, split into two primary tasks,
for improved reusability and testability.  This change works toward the larger
goal of making the code more maintainable.

The use of a simple main function, built with the modulino pattern in Perl,
allows the other subroutines to be run by unit tests.  Compared to acceptance
tests used until now, which require an exec call, this enables unit test to
run in the same process, preserving stack information on exceptions.  This
empowers developers to create faster, and potentially more informative, tests.